Rename websocket package
ws -> websocket
ws/websocket -> websocket/server
