+++ websocket/session/session.go
